feat: Reasonable default domain for tunnelhost
People installing the helm chart into another namespace then we usually
document may experience hidden issues.
The current default values are very opiniated and assume the helm chart
will be installed in "hub-agent".

In a default kubernetes cluster the dns resolution is usually able to leverage the
search domain to reach services in the same namespace without knowing the
fqdn.